+++ src/domains/kebab-case/index.ts
